Split project and upgrade gradle
