+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="net.ahammad.showhiddenpassword" >
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    package="net.ahammad.showhiddenpassword.app" >
 
     <application
-        android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
-        <activity
-            android:name="net.ahammad.showhiddenpassword.MyActivity"
-            android:label="@string/app_name" >
+        <activity android:name=".MyActivity" android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
