+++ library/src/main/res/layout/view_edittext_showpassword.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent" android:layout_height="wrap_content">
+
+    <EditText
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/edittext"
+        android:password="true"
+        android:paddingRight="35dip"/>
+
+    <Button
+        android:id="@+id/button_clear"
+        android:layout_width="25dip"
+        android:layout_height="20dip"
+        android:layout_alignParentRight="true"
+        android:background="@drawable/show_pass"
+        android:layout_centerVertical="true"
+        android:layout_marginRight="5dip" />
+</RelativeLayout>
