+++ library/src/main/res/values/attrs.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="ShownEdittext">
+        <attr name="android:inputType"/>
+        <attr name="android:hint"/>
+        <attr name="android:enabled"/>
+    </declare-styleable>
+</resources>
