+++ settings.gradle
@@ -1,1 +1,2 @@
+include ':library'
 include ':app'
